Complete Week 3 Implementation: Feature-based restructure with barrel exports and documentation
