+++ src/features/search/components/SearchResults.jsx
@@ -1,7 +1,6 @@
 /* eslint-disable react/prop-types */
-import Button from '../../../components/ui/Button.jsx';
-import Badge from '../../../components/ui/Badge.jsx';
-import { toTitleCase } from '../../../utils/textUtils.js';
+import { Button, Badge } from '../../../components/ui';
+import { toTitleCase } from '../../../utils';
 
 /**
  * SearchResults Component
